refactor (reduced code duplication)
